+++ packages/divider/src/Divider.stories.tsx
@@ -3,6 +3,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { Divider } from './Divider';
 
 const meta = {
+  title: 'Divider',
   component: Divider,
   parameters: {
     layout: 'centered',