+++ packages/radio-group/.storybook/preview.ts
@@ -1,0 +1,5 @@
+import type { Preview } from '@storybook/react';
+
+export default {
+  tags: ['autodocs'],
+} satisfies Preview;