+++ packages/typography/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @sipe-team/typography
 
+## 0.0.4-next.1
+
+### Patch Changes
+
+- fix: exports
+  - @sipe-team/tokens@0.1.0
+
+## 0.0.4-next.0
+
+### Patch Changes
+
+- 27c312a: fix: styles.css
+  - @sipe-team/tokens@0.1.0
+
+## 0.0.3
+
+### Patch Changes
+
+- e6f76c0: fix(typography): css module bundle
+  - @sipe-team/tokens@0.1.0
+
+## 0.0.3-next.0
+
+### Patch Changes
+
+- 1885991: fix(typography): css module bundle
+  - @sipe-team/tokens@0.1.0
+
+## 0.0.2
+
+### Patch Changes
+
+- 9c93399: feat(tokens): add tokens package
+  fix(typography): reuse `@sipe-team/tokens`
+- Updated dependencies [9c93399]
+  - @sipe-team/tokens@0.1.0
+
 ## 0.0.1
 
 ### Patch Changes